Limit how many bytes are copied from the command line.
The new option `TIMEHISTORY_CMDLINE_LIMIT` (512 by default) can be used to limit
how many bytes are copied from the command line to the history entry.

If a command line exceeds this limit, the arguments are truncated.

The size of a CString written to the shared buffer is written as a `usize`
instead of `u16`.